Rename webpack-plugins to plugins and WebpackPluginsUtil to FrontendPluginsUtil
