+++ flow-server/src/main/resources/plugins/plugins.json
